requirements.txt
torch
torchvision
gradio
numpy
pillow